+++ extra_features/influxdb_adapter/data_collection_queries/avg_sproc_runtime_per_schema.sql
@@ -1,6 +1,6 @@
 select
   sp_timestamp as "timestamp",
-  extract(epoch from sp_timestamp::timestamp with time zone at time zone 'utc') as "time",
+  extract(epoch from sp_timestamp::timestamp with time zone at time zone 'utc')::int as "time",
   sproc_schema as "schema",
   ((calls_delta / extract (epoch from timestamp_delta)) * 3600) ::int8 as calls_1h_rate,
   (self_delta / calls_delta * 1000)::int8 as avg_self_us,
