+++ extra_features/influxdb_adapter/data_collection_queries/sproc_details_per_schema_sproc.sql
@@ -1,6 +1,6 @@
 select
   sp_timestamp as "timestamp",
-  extract(epoch from sp_timestamp::timestamp with time zone at time zone 'utc') as "time",
+  extract(epoch from sp_timestamp::timestamp with time zone at time zone 'utc')::int as "time",
   sproc_schema as "schema",
   sproc,
   ((calls_delta / timestamp_delta_s) * 3600) ::int8 as calls_1h_rate,
